backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
dashscope==1.25.16
python-dotenv==1.1.0
websockets==15.0.1
